+++ src/main/java/firok/tiths/common/Items.java
@@ -67,6 +67,8 @@ public class Items
 	public static final Item moonStone = new Item(); // 月光石
 	@Reg(nameTreeRoot)
 	public static final Item treeRoot = new Item(); // 树根
+	@Reg(nameBrokenBedrock)
+	public static final Item brokenBedrock = new Item(); // 破碎基岩
 
 	// 金属粒
 	@Reg(Keys.nuggetStellarium)